Move db module to adapters package
