+++ app/api/services/users/create_user_csv.py
@@ -5,7 +5,7 @@ from dataclasses import asdict, dataclass
 
 from smart_open import open as smart_open
 
-import api.db as db
+import api.adapters.db as db
 import api.logging
 from api.db.models.user_models import User
 from api.scripts.util.script_util import script_context_manager
