+++ app/tests/api/db/test_db.py
@@ -1,12 +1,7 @@
 import logging  # noqa: B1
 
-from api.db import (
-    DbConfig,
-    get_connection_parameters,
-    get_db_config,
-    make_connection_uri,
-    verify_ssl,
-)
+from api.adapters.db.client import get_connection_parameters, make_connection_uri, verify_ssl
+from api.adapters.db.config import DbConfig, get_db_config
 
 
 class DummyConnectionInfo:
