+++ app/tests/conftest.py
@@ -6,8 +6,8 @@ import moto
 import pytest
 import sqlalchemy
 
+import api.adapters.db as db
 import api.app as app_entry
-import api.db as db
 import api.logging
 import tests.api.db.models.factories as factories
 from api.db import models
