+++ docs/app/database/database-access-management.md
@@ -16,7 +16,7 @@ For example, **do this**
 ### right way ###
 
 from flask import current_app
-import api.db as db
+import api.adapters.db as db
 
 def some_service_func(session: db.Session)
     with db_session.begin(): # start transaction
@@ -35,7 +35,7 @@ and **don't do this**
 ### wrong way ###
 
 from flask import current_app
-import api.db as db
+import api.adapters.db as db
 
 def some_service_func()
     db_client = db.get_db(current_app)
